+++ listaAdjacencia.h
@@ -9,17 +9,17 @@ typedef struct listaAdjacencia ListaAdjacencia;
 
 typedef struct grafo Grafo;
 
-Node* criaNode(int origem, int destino, float distancia, float velocidade);
+Node* criaNode(int origem, int destino, double distancia, double velocidade);
 
 Grafo* criaGrafo(int vertices);
 
-void insereAresta(Grafo* grafo, int origem, int destino, float distancia, float velocidade);
+void insereAresta(Grafo* grafo, int origem, int destino, double distancia, double velocidade);
 
 Grafo* initGrafo(FILE* arquivoEntrada, int* nVertices, int* mArestas, int* vOrigem, int* vDestino, Atualizacoes* atualizacoes);
 
 void liberaGrafo(Grafo* grafo);
 
-void atualizaGrafo(Grafo** grafo, int origem, int destino, float novaVelocidade);
+void atualizaGrafo(Grafo** grafo, int origem, int destino, double novaVelocidade);
 
 int getVertices(Grafo* grafo);
 
@@ -29,9 +29,9 @@ Node* getProx(Node* node);
 
 int getDestino(Node* node);
 
-float getDistancia(Node* node);
+double getDistancia(Node* node);
 
-float getTempo(Node* node);
+double getTempo(Node* node);
 
 int getOrigem(Node* node);
